renames helpers to utils and moves files inside
