+++ lib/middleware.js
@@ -1,10 +1,10 @@
 "use strict";
 
-const cryptoStr = require("../helper/hash");
+const cryptoStr = require("../utils/hash");
 const postgres = require("./postgres-db");
 const mongo = require("./mongo-db");
-const parseRequest = require("./parse-request");
-const extractPayload = require("./extract-payload");
+const parseRequest = require("../utils/parse-request");
+const extractPayload = require("../utils/extract-payload");
 const { wss } = require("../bin/www/server");
 const { WebSocket } = require("ws");
