+++ lib/postgres-db.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const { dbQuery } = require("../helper/db-query");
+const { dbQuery } = require("../utils/db-query");
 
 const generateTemporaryEndPointName = (() => {
   let count = 0;
