+++ utils/hash.js
